Double auto almost done!

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Components/Drivetrain/TurnTo.java
@@ -210,7 +210,7 @@ public class TurnTo extends RobotCommand
     }
 
 
-    public void runSequentially(int BUFFER_PARAM, double SECONDS_TIMEOUT)
+    public void runSequentially(double BUFFER_PARAM, double SECONDS_TIMEOUT)
     {
         TIMEOUT = (long)(SECONDS_TIMEOUT * 1000);
         if(_drivetrain.getEncoderMode() != DcMotor.RunMode.RUN_USING_ENCODER)
